Refactor VideoTube detail view helpers
Manual testing: not run (automated test coverage only).
