+++ src/ui/views/browser/components/videotube/views/detail/index.js
@@ -1,0 +1,6 @@
+export { renderRenameForm } from './renameForm.js';
+export { renderStatsGrid } from './statsGrid.js';
+export { renderQualityPanel } from './qualityPanel.js';
+export { renderPayoutPanel } from './payoutPanel.js';
+export { renderNichePanel } from './nichePanel.js';
+export { renderActionsPanel } from './actionsPanel.js';
